backend/no-server/requirements.txt
Flask==2.2.2
sanic==19.6.0
python-dotenv==0.14.0
pymongo==4.3.3
uniswap-python==0.6.2
